engine_python/main.py
import os
import time
import numpy as np
import polars as pl
import grpc
import redis
import mxnet as mx
from mxnet import nd, gluon

# =================================================================
# BLOCO DE COMPLIANCE, NORMAS E GRAMÁTICA (Sincronizados)
# =================================================================
try:
    from utils.compliance import NormaABNT
    from utils.meta_expert import MetaExpert
    from utils.gramatica_expert import GramaticaExpert
except ImportError:
    import sys
    sys.path.append('/app')
    from utils.compliance import NormaABNT
    from utils.meta_expert import MetaExpert
    from utils.gramatica_expert import GramaticaExpert

from utils.memoria_zeta import MemoriaZeta

# Bloqueio de Compatibilidade NumPy/MXNet
compat_map = {'bool': bool, 'float': float, 'int': int, 'object': object}
for attr, val in compat_map.items():
    if not hasattr(np, attr): setattr(np, attr, val)

# Conexão com o Pulso de Memória (Redis) para Simulação Dinâmica
try:
    r_cache = redis.Redis(host='redis', port=6379, db=0, decode_responses=True)
except:
    r_cache = None

def executar_instrumentacao_zeta():
    print("\n" + "="*80)
    print("🔬 MONITORAMENTO DE MICROESTRUTURA: MODO SIMULAÇÃO ATIVO")
    print("⚔️  AUDITORIA: ABNT-ZETA-1200 | REATOR DE ISÓTOPOS (Ruby-Python)")
    print("="*80 + "\n")

    # Instanciamos o Grampeador de Fase (SentenceTransformers)
    auditor_gramatical = GramaticaExpert()
    
    net = gluon.nn.Sequential() 
    net.initialize(ctx=mx.cpu())

    while True:
        try:
            # 1. ESCUTA DE ISÓTOPO (Sinal vindo do Ruby via Redis)
            v_alvo = r_cache.get("manifold:target_zeta") if r_cache else None
            
            # 2. EVOCAÇÃO DO GRADIENTE (Base ou Injetado)
            massa_base = float(v_alvo) if v_alvo else MetaExpert.calcular_funcao_limite()
            
            # 3. IDENTIFICAÇÃO DE FENOMENOLOGIA (Busca a Memória Zeta)
            assinatura, meta = MemoriaZeta.evocar_memoria(massa_base)
            
            # 4. ENTRADA DE DIÁLOGO DINÂMICA
            if assinatura == "PIRIMIDINA":
                frase_input = "O Isótopo Orto oscila no Manifold Para sob a Tensão do Nitrogênio."
            else:
                frase_input = "O Isótopo Orto, resvala no Manifold Para, sob a Égide de Hückel."
            
            # 5. AUDITORIA DE CRUZAMENTO (Gramática Física + Geometria)
            # Captura de 5 variáveis para evitar erro de unpack
            status, conf, veredito, v_zeta, equacao = auditor_gramatical.validar_sintaxe_zeta(
                frase_input, 
                massa_base, 
                assinatura
            )

            # 6. TELEMETRIA TÉCNICA (Veredito de Simulação)
            print(f"[INSTRUMENTAÇÃO DINÂMICA T={time.time():.4f}]")
            print(f"| ISÓTOPO_INJETADO: {assinatura}")
            print(f"| EQUAÇÃO_ESTADO: {equacao}")
            print(f"| CONJUNÇÃO_LÓGICA: {veredito}")
            print(f"| GRADIENTE_RICCI_OBSERVADO: {v_zeta:.10f}")

            # Auditoria ABNT
            is_conforme, log_compliance = NormaABNT.auditar_conformidade_tecnica(v_zeta)

            if is_conforme and status == "ESTÁVEL":
                print(f"| STATUS_COMPLIANCE: {log_compliance}")
                print("| DIAGNÓSTICO: CONVERGÊNCIA_ADIABÁTICA_CONFIRMADA")
            elif status == "DERIVA":
                print(f"| STATUS_COMPLIANCE: ALERTA_DE_DERIVA")
                print("| DIAGNÓSTICO: RESSONÂNCIA_HETEROCÍCLICA_ATIVA (1+1 ≈ Δ)")
            else:
                print(f"| STATUS_COMPLIANCE: {log_compliance}")
                print("| DIAGNÓSTICO: RUPTURA_DE_GRADIENTE (ASSIMETRIA_DETECTADA)")

            print("-" * 80)
            time.sleep(5)

        except Exception as e:
            print(f"🚩 FALHA NO REATOR: {e}")
            time.sleep(2)

if __name__ == "__main__":
    executar_instrumentacao_zeta()

